refactor: simplify graceful shutdown configuration
Replace the profile and builder APIs with one source-consumable service collection extension. Remove the premature test project and its dedicated workflow.
