--- .github/workflows/common-dotnet-tests.yaml
@@ -1,45 +1,0 @@
-name: Common .NET - Build and test
-
-on:
-  pull_request:
-    types: [opened, reopened, synchronize, ready_for_review, converted_to_draft, closed]
-    paths:
-      - 'src/common/dotnet/Altinn.Studio.Common/**'
-      - '.github/workflows/common-dotnet-tests.yaml'
-      - '!**/AGENTS.md'
-      - '!**/CLAUDE.md'
-  workflow_dispatch:
-
-concurrency:
-  group: ${{ github.workflow }}-${{ github.event.pull_request.number || github.ref || github.run_id }}
-  cancel-in-progress: true
-
-jobs:
-  build-and-test:
-    if: ${{ github.event_name != 'pull_request' || (github.event.action != 'closed' && !github.event.pull_request.draft) }}
-    name: Build and test
-    runs-on: ubuntu-latest
-    timeout-minutes: 10
-    defaults:
-      run:
-        working-directory: src/common/dotnet/Altinn.Studio.Common
-    steps:
-      - name: Checkout code
-        uses: actions/checkout@3d3c42e5aac5ba805825da76410c181273ba90b1 # v7.0.1
-
-      - name: Setup .NET
-        uses: actions/setup-dotnet@26b0ec14cb23fa6904739307f278c14f94c95bf1 # v5.4.0
-        with:
-          dotnet-version: 10.0.x
-
-      - name: Restore tools
-        run: dotnet tool restore
-
-      - name: Check formatting
-        run: dotnet csharpier check .
-
-      - name: Build
-        run: dotnet build Altinn.Studio.Common.slnx
-
-      - name: Test
-        run: dotnet test Altinn.Studio.Common.slnx --no-build
