+++ src/Runtime/gateway/src/Altinn.Studio.Gateway.Api/Program.cs
@@ -9,11 +9,10 @@ var builder = WebApplication.CreateSlimBuilder(args);
 builder.AddGateway();
 builder.ConfigureKestrelPorts(publicPort: PortConfiguration.PublicPort, internalPort: PortConfiguration.InternalPort);
 builder.UseCommonHostingConfiguration();
-builder.UseGracefulShutdown(
-    new GracefulShutdownProfile(
-        endpointDrainDelay: TimeSpan.FromSeconds(5),
-        applicationShutdownTimeout: TimeSpan.FromSeconds(20)
-    )
+builder.Services.AddGracefulShutdown(
+    builder.Environment,
+    endpointDrainDelay: TimeSpan.FromSeconds(5),
+    applicationShutdownTimeout: TimeSpan.FromSeconds(20)
 );
 builder.AddMaskinportenAuthentication();
 builder.AddOpenTelemetry();
