--- src/common/dotnet/Altinn.Studio.Common/Altinn.Studio.Common.slnx
@@ -1,4 +1,0 @@
-<Solution>
-  <Project Path="Altinn.Studio.Common.csproj" />
-  <Project Path="tests/Altinn.Studio.Common.Tests/Altinn.Studio.Common.Tests.csproj" />
-</Solution>
